Create queue operation
Array Initialization:

The queue is represented using an array (self.queue), initialized with None values of a fixed size (size).
is_full Method:

This method checks if the queue is full by comparing the rear index with max_size - 1.
is_empty Method:

This method checks if the queue is empty by checking if the front index is -1 or if front > rear.
enqueue Method:

Adds an item to the rear of the queue. If it's the first item, the front is set to 0.
If the queue is full, it displays a message.
dequeue Method:

Removes an item from the front of the queue and increments the front index.
If the queue becomes empty after a dequeue, both front and rear are reset to -1.
display Method:

Displays the current elements in the queue from front to rear.